Added Home Page For Heads
Created a home page for heads which enables the head to view the current
status of all the sub heads.

Also the form to assign a task to a sub head has been added. It will be
included as a popup in the final version of the page. This popup will be
available only for the subheads with no active assignment.

Fixed an error in class-task. The assignTask function now sets the
engage field of the subhead to true.