Fixed Assign Tasks and Edit Tasks
The heads can now assign and edit tasks for single subhead by clicking
on their cards and entering the necessary data into the popups.

The generation of cards has been made dynamic and one is generated for
each subhead, with the ones having a task previously allocated to them
first.

There was a bug in the previous version which cause the data of the
first field to get updated. It has now been fixed and the correct fields
are updated.

Materialize date picker has been removed because it took too much space.